requirements.txt
aniso8601==8.0.0
attrs==19.3.0
cachetools==3.1.1
certifi==2019.6.16
chardet==3.0.4
Click==7.0
Flask==1.1.1
flask-restplus==0.13.0
google-api-core==1.14.2
google-auth==1.6.3
google-cloud-core==1.0.3
google-cloud-datastore==1.9.0
google-cloud-storage==1.18.0
google-resumable-media==0.3.2
googleapis-common-protos==1.6.0
grpcio==1.22.0
idna==2.8
importlib-metadata==0.23
itsdangerous==1.1.0
Jinja2==2.10.3
jsonschema==3.1.1
MarkupSafe==1.1.1
more-itertools==7.2.0
numpy==1.17.0
pika==1.1.0
protobuf==3.9.1
protobuf3==0.2.1
pyasn1==0.4.6
pyasn1-modules==0.2.6
pyrsistent==0.15.5
pytz==2019.2
requests==2.22.0
rsa==4.0
six==1.12.0
tiingo==0.10.0
urllib3==1.25.3
Werkzeug==0.16.0
zipp==0.6.0
